Created JS structure
