+++ source/static/js/attribute-menu.js
@@ -1,0 +1,9 @@
+export class AttributeMenu {
+    constructor() {
+        this.obj = null;
+    }
+
+    updateObject(obj) {
+        this.obj = obj;
+    }
+}
